+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0]
+
+- Add sort and seller options to getResaleItems
+- Add useAuth, account option to get user showcases
+- Add file and description to create showcase
+- Update modify showcase
+- Add get all collectibles for user
+- Add gem methods
+
 ## [3.2.4]
 
 - Adding Gem staking methods